+++ .travis.yml
@@ -1,0 +1,7 @@
+language: node_js
+node_js:
+  - "9"
+
+script:
+  - npm run lint
+  - npm test
